+++ src/main/java/com/example/bilda_server/domain/entity/Team.java
@@ -1,4 +1,4 @@
-package com.example.bilda_server.domain;
+package com.example.bilda_server.domain.entity;
 
 import com.example.bilda_server.domain.enums.CompleteStatus;
 import com.example.bilda_server.domain.enums.RecruitmentStatus;
